Fixing project.json generation issue...
The project json generated was faulty because the insertion of template 10 nugget into json was inserted at position based on newtonsoft - which had a version update therefore was no longer correct. This approach is not impacted by version updates like that. We need this code in order to convert from the project reference approach in the project based libraries and our final installer which uses nugget package references instead.